+++ example/ios/Runner/AppDelegate.swift
@@ -35,7 +35,7 @@ import flutter_callkit_incoming
         let nameCaller = handleObj.getDecryptHandle()["nameCaller"] as? String ?? ""
         let handle = handleObj.getDecryptHandle()["handle"] as? String ?? ""
         let data = flutter_callkit_incoming.Data(id: UUID().uuidString, nameCaller: nameCaller, handle: handle, type: isVideo ? 1 : 0)
-        //set more data
+        //set more data...
         data.nameCaller = "Johnny"
         SwiftFlutterCallkitIncomingPlugin.sharedInstance?.startCall(data, fromPushKit: true)